fix: move options spread to first position
When composing an object, entries to the right take precedence over entries to the left. Overwrites should thus be positioned to the right of the `...options` spread. This actually causes a bug on line 472 where the presentation `purpose` is overwritten by any `purpose` property in the options (which is meant to be passed down to `verifyCredential` instead).